+++ app/src-tauri/src/pkce.rs
@@ -10,10 +10,7 @@ use actix_web::{
     web::{self, Data, Query},
     App, HttpRequest, HttpResponse, HttpServer, Result,
 };
-use apicize_lib::{
-    oauth2_client_tokens::PkceTokenResult,
-    oauth2_pkce::{generate_authorization, refresh_token, retrieve_access_token},
-};
+use apicize_lib::{oauth2_pkce::{generate_authorization, refresh_token, retrieve_access_token}, PkceTokenResult};
 use serde::{Deserialize, Serialize};
 use tauri::{AppHandle, Emitter, Manager, Url};